# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/products/barcode/{barcode}": {
      "get": {
        "operationId": "ProductsController_findByBarcode",
        "summary": "Buscar produto por código de barras",
        "parameters": [
          {
            "name": "barcode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado com sucesso"
          },
          "404": {
            "description": "Produto não encontrado"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/seller/sales": {
      "post": {
        "operationId": "SellerController_createSale",
        "summary": "Criar venda no PDV",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Venda criada com sucesso"
          },
          "400": {
            "description": "Dados inválidos ou estoque insuficiente"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/sales/history": {
      "get": {
        "operationId": "SellerController_getSaleHistory",
        "summary": "Histórico de vendas do vendedor",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico retornado com sucesso"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/sales/suspended": {
      "get": {
        "operationId": "SellerController_getSuspendedSales",
        "summary": "Listar vendas suspensas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Vendas suspensas retornadas com sucesso"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/sales/{id}": {
      "get": {
        "operationId": "SellerController_getSaleById",
        "summary": "Detalhes de uma venda",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Venda retornada com sucesso"
          },
          "404": {
            "description": "Venda não encontrada"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/sales/{id}/suspend": {
      "post": {
        "operationId": "SellerController_suspendSale",
        "summary": "Suspender uma venda",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Venda suspensa com sucesso"
          },
          "404": {
            "description": "Venda não encontrada"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/sales/{id}/resume": {
      "post": {
        "operationId": "SellerController_resumeSale",
        "summary": "Retomar uma venda suspensa",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Venda retomada com sucesso"
          },
          "404": {
            "description": "Venda não encontrada"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/stats": {
      "get": {
        "operationId": "SellerController_getSellerStats",
        "summary": "Estatísticas do vendedor (hoje)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Estatísticas retornadas com sucesso"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/orders/pending": {
      "get": {
        "operationId": "SellerController_getPendingOrders",
        "summary": "Pedidos pendentes para picking",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Pedidos pendentes retornados com sucesso"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/orders/{orderId}/accept": {
      "post": {
        "operationId": "SellerController_acceptOrder",
        "summary": "Aceitar pedido para picking",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido aceito com sucesso"
          },
          "404": {
            "description": "Pedido não encontrado"
          },
          "409": {
            "description": "Pedido já aceito por outro vendedor"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/picking": {
      "get": {
        "operationId": "SellerController_getMyPickingOrders",
        "summary": "Minhas ordens de picking ativas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Ordens retornadas com sucesso"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/picking/{id}": {
      "get": {
        "operationId": "SellerController_getPickingOrder",
        "summary": "Detalhes de ordem de picking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ordem retornada com sucesso"
          },
          "404": {
            "description": "Ordem não encontrada"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/picking/{pickingOrderId}/scan": {
      "post": {
        "operationId": "SellerController_scanItem",
        "summary": "Escanear item no picking",
        "parameters": [
          {
            "name": "pickingOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item escaneado"
          },
          "404": {
            "description": "Ordem de picking não encontrada"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/picking/{pickingItemId}/manual-pick": {
      "post": {
        "operationId": "SellerController_markItemPicked",
        "summary": "Marcar item como coletado manualmente",
        "parameters": [
          {
            "name": "pickingItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item marcado como coletado"
          },
          "404": {
            "description": "Item não encontrado"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/picking/{pickingOrderId}/complete": {
      "post": {
        "operationId": "SellerController_completePickingOrder",
        "summary": "Concluir ordem de picking",
        "parameters": [
          {
            "name": "pickingOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ordem concluída com sucesso"
          },
          "404": {
            "description": "Ordem não encontrada"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/products/barcode/{barcode}": {
      "get": {
        "operationId": "SellerController_getProductByBarcode",
        "summary": "Buscar produto por código de barras",
        "parameters": [
          {
            "name": "barcode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado"
          },
          "404": {
            "description": "Produto não encontrado"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/customers/search": {
      "get": {
        "operationId": "SellerController_searchCustomers",
        "summary": "Buscar clientes",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Termo de busca",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clientes encontrados"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/seller/customers/quick": {
      "post": {
        "operationId": "SellerController_quickCreateCustomer",
        "summary": "Criar cliente rápido",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente criado com sucesso"
          },
          "409": {
            "description": "Telefone já cadastrado"
          }
        },
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Supermercado API",
    "description": "API completa para plataforma de supermercado - Venâncio Aires, RS",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SaleItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "description": "ID do produto"
          },
          "quantity": {
            "type": "number",
            "description": "Quantidade",
            "minimum": 1
          },
          "unitPrice": {
            "type": "number",
            "description": "Preço unitário"
          },
          "discount": {
            "type": "number",
            "description": "Desconto do item"
          }
        },
        "required": [
          "productId",
          "quantity",
          "unitPrice"
        ]
      },
      "CreateSaleDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Itens da venda",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaleItemDto"
            }
          },
          "customerId": {
            "type": "string",
            "description": "ID do cliente"
          },
          "discount": {
            "type": "number",
            "description": "Desconto geral da venda"
          },
          "paymentMethod": {
            "type": "string",
            "description": "Método de pagamento",
            "enum": [
              "CASH",
              "CREDIT_CARD",
              "DEBIT_CARD",
              "PIX",
              "TRANSFER"
            ]
          },
          "paidAmount": {
            "type": "number",
            "description": "Valor pago (para cálculo de troco)"
          },
          "notes": {
            "type": "string",
            "description": "Observações da venda"
          }
        },
        "required": [
          "items",
          "paymentMethod"
        ]
      },
      "ScanItemDto": {
        "type": "object",
        "properties": {
          "barcode": {
            "type": "string",
            "description": "Código de barras do produto"
          }
        },
        "required": [
          "barcode"
        ]
      },
      "QuickCreateCustomerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do cliente"
          },
          "phone": {
            "type": "string",
            "description": "Telefone do cliente"
          },
          "email": {
            "type": "string",
            "description": "E-mail do cliente"
          }
        },
        "required": [
          "name",
          "phone"
        ]
      },
...
    }
  }
}